+++ resources/views/partials/config.blade.php
@@ -1,3 +1,0 @@
-@if ($timeZone)
-<script>document.documentElement.setAttribute('time-zone', @json($timeZone));</script>
-@endif